[refactor] 모집 글 목록, 리뷰 글 목록, 커뮤니티 글 목록 리팩토링
