+++ .eslintrc
@@ -74,6 +74,6 @@
     ],
   },
   "env": {
-    "browser": true, // document, window 등의 브라우저 내장 객체들 조회시 ESLint 오류 없이 작업
+    "browser": true,
   },
 }
