+++ src/App.css
@@ -1,3 +1,11 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+.post-flex-container {
+	@apply flex flex-wrap gap-[35px];
+}
+
+.post-grid-container {
+	@apply grid grid-cols-2 gap-[20px] w-full min-w-[700px];
+}
