--- src/Page/Home.tsx
@@ -1,5 +1,0 @@
-const Home = (): JSX.Element => {
-  return <div></div>;
-};
-
-export default Home;
